refactor: split input field types into dedicated files
Follow existing one-type-per-file pattern (ticket-extra-option.ts,
opening-hours-time.ts, etc.) by splitting the monolithic
input-field-creation-request.ts into 16 individual files.
