+++ packages/javascript-api/src/lib/model/input-field/date-field-creation-request.ts
@@ -1,0 +1,9 @@
+import { InputFieldBase } from './input-field-base.js';
+import { InputFieldTranslation } from './input-field-translation.js';
+
+export interface DateFieldCreationRequest extends InputFieldBase {
+  type: 'DATE';
+  title: string;
+  visitorFacingTitle?: string;
+  translations?: InputFieldTranslation[];
+}
