+++ packages/javascript-api/src/lib/model/input-field/email-field-creation-request.ts
@@ -1,0 +1,6 @@
+import { InputFieldBase } from './input-field-base.js';
+
+export interface EmailFieldCreationRequest extends InputFieldBase {
+  type: 'EMAIL';
+  isRequiredInAppointments: boolean;
+}
