+++ packages/javascript-api/src/lib/model/input-field/first-name-field-creation-request.ts
@@ -1,0 +1,5 @@
+import { InputFieldBase } from './input-field-base.js';
+
+export interface FirstNameFieldCreationRequest extends InputFieldBase {
+  type: 'FIRST_NAME';
+}
