+++ packages/javascript-api/src/lib/model/input-field/input-field-translation.ts
@@ -1,0 +1,5 @@
+export interface InputFieldTranslation {
+  languageCode: string;
+  title?: string;
+  visitorFacingTitle?: string;
+}
