+++ packages/javascript-api/src/lib/model/input-field/input-field-type.ts
@@ -1,0 +1,10 @@
+export type InputFieldType =
+  | 'TEXT'
+  | 'SELECT'
+  | 'EMAIL'
+  | 'PHONE_NUMBER'
+  | 'FIRST_NAME'
+  | 'LAST_NAME'
+  | 'URL'
+  | 'DATE'
+  | 'NUMERIC';
