+++ packages/javascript-api/src/lib/model/input-field/last-name-field-creation-request.ts
@@ -1,0 +1,6 @@
+import { InputFieldBase } from './input-field-base.js';
+
+export interface LastNameFieldCreationRequest extends InputFieldBase {
+  type: 'LAST_NAME';
+  isRequiredInAppointments?: boolean;
+}
