+++ packages/javascript-api/src/lib/model/input-field/numeric-field-constraints.ts
@@ -1,0 +1,5 @@
+export interface NumericFieldConstraints {
+  min?: number;
+  max?: number;
+  scale: number;
+}
