+++ packages/javascript-api/src/lib/model/input-field/numeric-field-creation-request.ts
@@ -1,0 +1,11 @@
+import { InputFieldBase } from './input-field-base.js';
+import { InputFieldTranslation } from './input-field-translation.js';
+import { NumericFieldConstraints } from './numeric-field-constraints.js';
+
+export interface NumericFieldCreationRequest extends InputFieldBase {
+  type: 'NUMERIC';
+  title: string;
+  visitorFacingTitle?: string;
+  translations?: InputFieldTranslation[];
+  constraints?: NumericFieldConstraints;
+}
