+++ packages/javascript-api/src/lib/model/input-field/phone-number-field-creation-request.ts
@@ -1,0 +1,5 @@
+import { InputFieldBase } from './input-field-base.js';
+
+export interface PhoneNumberFieldCreationRequest extends InputFieldBase {
+  type: 'PHONE_NUMBER';
+}
