+++ packages/javascript-api/src/lib/model/input-field/select-option-translation.ts
@@ -1,0 +1,4 @@
+export interface SelectOptionTranslation {
+  languageCode: string;
+  title?: string;
+}
