+++ packages/javascript-api/src/lib/model/input-field/select-option.ts
@@ -1,0 +1,10 @@
+import { UUID } from '../uuid.js';
+import { SelectOptionTranslation } from './select-option-translation.js';
+
+export interface SelectOption {
+  /** Client-generated UUID for this option. */
+  id: UUID;
+  title: string;
+  color?: string;
+  translations?: SelectOptionTranslation[];
+}
