+++ packages/javascript-api/src/lib/model/input-field/url-field-creation-request.ts
@@ -1,0 +1,8 @@
+import { InputFieldBase } from './input-field-base.js';
+import { InputFieldTranslation } from './input-field-translation.js';
+
+export interface UrlFieldCreationRequest extends InputFieldBase {
+  type: 'URL';
+  title: string;
+  translations?: InputFieldTranslation[];
+}
